adapter: take over notice severity checking
Previously this was in pgwire, so it didn't work for other adapter
client users like websocket. Move checking into the adapter so no client
has to be aware of this.

Fixes https://github.com/MaterializeInc/console/issues/681
