+++ src/pgwire/src/lib.rs
@@ -95,7 +95,6 @@ mod protocol;
 mod server;
 
 pub use codec::MAX_REQUEST_SIZE;
-pub use message::Severity;
 pub use metrics::MetricsConfig;
 pub use protocol::match_handshake;
 pub use server::{Config, Server, TlsConfig, TlsMode};
